refactor(web): move target label to form, rename hook result type
